inclusão da imagem na tela de login
