--- Circulando/Controller/LoginView.swift
@@ -1,20 +1,0 @@
-//
-//  LoginView.swift
-//  Circulando
-//
-//  Created by Izadora Netz on 28/11/22.
-//
-
-import Foundation
-import UIKit
-
-class LoginView: UIView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.backgroundColor = .white
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
