+++ Circulando/Resource/Assets.xcassets/ImageLogin.imageset/Contents.json
@@ -1,0 +1,24 @@
+{
+    "images" : [
+      {
+        "filename" : "ImageLogin.png",
+        "idiom" : "universal",
+        "scale" : "1x"
+      },
+      {
+        "filename" : "ImageLogin@2x.png",
+        "idiom" : "universal",
+        "scale" : "2x"
+      },
+      {
+        "filename" : "ImageLogin@3x.png",
+        "idiom" : "universal",
+        "scale" : "3x"
+      }
+    ],
+    "info" : {
+      "author" : "xcode",
+      "version" : 1
+    }
+  }
+
