+++ Circulando/Source/Controller/LoginView.swift
@@ -1,0 +1,45 @@
+//
+//  LoginView.swift
+//  Circulando
+//
+//  Created by Izadora Netz on 28/11/22.
+//
+
+import Foundation
+import UIKit
+
+class LoginView: UIView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        self.backgroundColor = .white
+
+        //inicializa o visual
+        setupVisualElements()
+    }
+
+    var imageLogin: UIImageView = {
+        let imagem = UIImageView ()
+        imagem.image = UIImage(named: "ImageLogin")
+        imagem.contentMode = .scaleAspectFit
+        imagem.translatesAutoresizingMaskIntoConstraints = false
+
+        return imagem
+    }()
+
+    func setupVisualElements() {
+        self.addSubview(imageLogin)
+
+        NSLayoutConstraint.activate([
+
+            imageLogin.widthAnchor.constraint(equalToConstant: 296),
+            imageLogin.heightAnchor.constraint(equalToConstant: 304.94),
+            imageLogin.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 47),
+            imageLogin.topAnchor.constraint(equalTo: self.topAnchor, constant: 179),
+
+        ])
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
